Move $getField pipeline tests to operator level
